adding Specimen to DatasetVersion
After giving it some thought (together with @olinux) it will give us more flexibility to add Specimen also directly to a DatasetVersion. 
It will on the one side allow the user a lower level of detail for registering data (that will of course not be our recommendation, but will lower the threshold for some), and on the other hand emphasizes the connection between specimen and dataset versions (which is also more intuitive for the curators I hope).