deleted IRI
based on a discussion with @olinux 

FileBundles are defined by their content (connected files) and do not necessarily have an IRI. 
For us files will connect to the FileBundles they belong to defining with that the zip of that Bundle. 
For reducing possible inconsistencies we therefore decided to delete IRI from FileBundle and only keep it for file repos and single files.