logic reassessment of file bundle groupings
The content of a file bundle is defined by a file path pattern.

The file path pattern defines the overall grouping mechanism (grouping type).
@olinux : there is a graph dependency here, because the grouping type of the file path pattern has to match the one of the file bundle.

Assuming that "study target" was selected as grouping type.
The old "descendedFrom" property does not allow any further specifications regarding study targets.

To include them, I changed the property name to "groupedBy" (to better fit sementically) and added the studyTarget as linked category. 

In addition, I changed the "groupedBy" property back to be an array. Based on how the file bundles are generated, they could be based on multiple entities of the same grouping type. This is particularly true for study targets (e.g., study target grouping could be based on specific area + cell type).
@olinux this includes again a graph validation: in "groupedBy" only entities are allowed of the specified grouping type.

To go even further (and becoming more flexible): the grouping type could be an array as well, allowing to create file bundles that could be grouped by e.g. a subject and a study target. 

@HumanBrainProject/openminds-developers what do you think? I will adopt this PR draft based on our discussion.